+++ xlive/H2MOD/Modules/Networking/CustomPackets/CustomPackets.cpp
@@ -1,3 +1,4 @@
+#include "stdafx.h"
 #include "CustomPackets.h"
 
 #include "H2MOD.h"
@@ -136,7 +137,7 @@ void __stdcall message_gateway_hook(void *thisx, network_address* addr, int mess
 	case e_network_message_types::team_change:
 	{
 		s_team_change* pak = (s_team_change*)packet;
-		h2mod->set_local_team_index(pak->team_index);
+		h2mod->set_local_team_index((BYTE)pak->team_index);
 		return;
 	}
