+++ xlive/H2MOD/Modules/RunLoop/RunLoop.cpp
@@ -1,8 +1,9 @@
-#include "H2MOD\Modules\RunLoop\RunLoop.h"
+#include "stdafx.h"
+#include "RunLoop.h"
+
 #include "H2MOD\Modules\Startup\Startup.h"
 #include "H2MOD\Modules\OnScreenDebug\OnscreenDebug.h"
 #include "H2MOD\Modules\Utils\Utils.h"
-#include "stdafx.h"
 #include <d3d9.h>
 #include <fstream>
 #include <winsock.h>
@@ -52,14 +53,14 @@ void hotkeyFuncAlignWindow() {
 	int interval_height = monitor_height / 2;
 	D3DVIEWPORT9 pViewport;
 	pDevice->GetViewport(&pViewport);
-	int width = interval_width * round(pViewport.Width / (double)interval_width);
-	int height = interval_height * round(pViewport.Height / (double)interval_height);
+	int width = (int)(interval_width * round(pViewport.Width / (double)interval_width));
+	int height = (int)(interval_height * round(pViewport.Height / (double)interval_height));
 	RECT gameWindowRect;
 	GetWindowRect(H2hWnd, &gameWindowRect);
 	int monitorXOffset = gameWindowRect.left - info.rcMonitor.left;
 	int monitorYOffset = gameWindowRect.top - info.rcMonitor.top;
-	int padX = interval_width * round(monitorXOffset / (double)interval_width);
-	int padY = interval_height * round(monitorYOffset / (double)interval_height);
+	int padX = (int)(interval_width * round(monitorXOffset / (double)interval_width));
+	int padY = (int)(interval_height * round(monitorYOffset / (double)interval_height));
 	int posX = info.rcMonitor.left + padX;
 	int posY = info.rcMonitor.top + padY;
